client step 1

+++ app/Http/Controllers/admin/RoleController.php
@@ -18,55 +18,67 @@ class RoleController extends Controller
             return $next($request);
         });
     }
-    function index()
+    function index(request $request)
     {
-        $roles = role::latest()->paginate(15);
-        $permissionParents = permission::where('parent_id', 0)->get();
-        return view('admin.role.index', compact('roles', 'permissionParents'));
+        if ($request->user()->cannot('view', role::class)) {
+            return redirect()->back()->with('notification', "Bạn không được phép truy cập");
+        } else {
+            $roles = role::latest()->paginate(15);
+            $permissionParents = permission::where('parent_id', 0)->get();
+            return view('admin.role.index', compact('roles', 'permissionParents'));
+        }
     }
 
     function add(request $request)
     {
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|max:225|unique:roles',
-        ], [
-            'required' => 'Không được để trống',
-            'max' => 'Độ dài vượt quá 225 ký tự',
-            'unique' => 'Đã tồn tại vai trò này'
-        ]);
-        if ($validator->fails()) {
-            $errors = $validator->errors()->toArray();
-            return response()->json(['code' => 0, 'errors' => $errors]);
+        if ($request->user()->cannot('create', role::class)) {
+            return response()->json(['code' => -1, 'msg' => 'Bạn không được phép thêm vai trò']);
         } else {
-            $role = role::create($request->all());
-            $role->permissions()->attach($request->permission_ids);
-            $roles = role::latest()->paginate(15);
-            $view = view('admin.role.main_data', compact('roles'))->render();
-            return response()->json(['msg' => 'Đã thêm vai trò', 'view' => $view]);
+            $validator = Validator::make($request->all(), [
+                'name' => 'required|max:225|unique:roles',
+            ], [
+                'required' => 'Không được để trống',
+                'max' => 'Độ dài vượt quá 225 ký tự',
+                'unique' => 'Đã tồn tại vai trò này'
+            ]);
+            if ($validator->fails()) {
+                $errors = $validator->errors()->toArray();
+                return response()->json(['code' => 0, 'errors' => $errors]);
+            } else {
+                $role = role::create($request->all());
+                $role->permissions()->attach($request->permission_ids);
+                $roles = role::latest()->paginate(15);
+                $view = view('admin.role.main_data', compact('roles'))->render();
+                return response()->json(['msg' => 'Đã thêm vai trò', 'view' => $view]);
+            }
         }
     }
 
     function update(request $request)
     {
-        $id = $request->id;
-        $validator = Validator::make($request->all(), [
-            'name' => "max:225|required|unique:roles,name,$id,id",
-        ], [
-            'required' => 'Không được để trống',
-            'unique' => 'đã tồn tại',
-            'max' => "Vượt quá ký tự quy đinh (225)"
-        ]);
-        if ($validator->fails()) {
-            $errors = $validator->errors()->toArray();
-            return response()->json(['code' => 0, 'errors' => $errors]);
+        if ($request->user()->cannot('update', role::class)) {
+            return response()->json(['msg' => 'Bạn không được sửa vai trò', 'code' => -1]);
         } else {
-            $role = role::find($id);
-            $role->update($request->all());
-            $permission_ids = [];
-            $role->permissions()->sync($request->permission_ids);
-            $roles = role::latest()->paginate(15);
-            $view = view('admin.role.main_data', compact('roles'))->render();
-            return response()->json(['msg' => 'Đã cập nhật vai trò', 'view' => $view]);
+            $id = $request->id;
+            $validator = Validator::make($request->all(), [
+                'name' => "max:225|required|unique:roles,name,$id,id",
+            ], [
+                'required' => 'Không được để trống',
+                'unique' => 'đã tồn tại',
+                'max' => "Vượt quá ký tự quy đinh (225)"
+            ]);
+            if ($validator->fails()) {
+                $errors = $validator->errors()->toArray();
+                return response()->json(['code' => 0, 'errors' => $errors]);
+            } else {
+                $role = role::find($id);
+                $role->update($request->all());
+                $permission_ids = [];
+                $role->permissions()->sync($request->permission_ids);
+                $roles = role::latest()->paginate(15);
+                $view = view('admin.role.main_data', compact('roles'))->render();
+                return response()->json(['msg' => 'Đã cập nhật vai trò', 'view' => $view]);
+            }
         }
     }
 
@@ -80,12 +92,17 @@ class RoleController extends Controller
         return response()->json(['role' => $role, 'viewPermission_data' => $viewPermission_data]);
     }
 
-    function delete(request $request) {
-        $role = role::find($request->id);
-        $role->permissions()->detach();
-        $role->delete();
-        $roles = role::latest()->paginate(15);
-        $view = view('admin.role.main_data', compact('roles'))->render();
-        return response()->json(['msg' => 'Đã xóa vai trò', 'view' => $view]);
+    function delete(request $request)
+    {
+        if ($request->user()->cannot('delete', role::class)) {
+            return response()->json(['msg' => 'Bạn không được xóa vai trò', 'code' => -1]);
+        } else {
+            $role = role::find($request->id);
+            $role->permissions()->detach();
+            $role->delete();
+            $roles = role::latest()->paginate(15);
+            $view = view('admin.role.main_data', compact('roles'))->render();
+            return response()->json(['msg' => 'Đã xóa vai trò', 'view' => $view]);
+        }
     }
 }
